requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pydantic
azure-cognitiveservices-speech
python-multipart
pillow
